Refactor project folder and Cmake structure for best practice
