+++ .gitignore
@@ -1,1 +1,3 @@
 build
+.vscode
+Binaries
