--- 3rd_party/google-test/CMakeLists.txt.in
@@ -1,19 +1,0 @@
-cmake_minimum_required(VERSION 3.0)
-
-project(googletest-download NONE)
-
-include(ExternalProject)
-
-# Version bfc0ffc8a698072c794ae7299db9cb6866f4c0bc happens to be master when I set this up.
-# To prevent an issue with accidentally installing GTest / GMock with your project you should use a
-# commit after 9469fb687d040b60c8749b7617fee4e77c7f6409
-# Note: This is after the release of v1.8
-ExternalProject_Add(googletest
-  URL               https://github.com/google/googletest/archive/bfc0ffc8a698072c794ae7299db9cb6866f4c0bc.tar.gz
-  SOURCE_DIR        "${CMAKE_CURRENT_BINARY_DIR}/googletest-src"
-  BINARY_DIR        "${CMAKE_CURRENT_BINARY_DIR}/googletest-build"
-  CONFIGURE_COMMAND ""
-  BUILD_COMMAND     ""
-  INSTALL_COMMAND   ""
-  TEST_COMMAND      ""
-)
