+++ CMakePresets.json
@@ -1,0 +1,55 @@
+{
+    "version": 2,
+    "configurePresets": [
+        {
+            "name": "Debug",
+            "displayName": "Debug Configuration",
+            "description": "Ninja build system is used for debugging.",
+            "generator": "Ninja",
+            "binaryDir": "${sourceDir}/Binaries/${presetName}/Intermediate",
+            "cacheVariables": {
+                "CMAKE_BUILD_TYPE": "Debug",
+                "CMAKE_TOOLCHAIN_FILE": "",
+                "CMAKE_INSTALL_PREFIX": "${sourceDir}/Binaries/${presetName}/Output",
+                "CMAKE_ARCHIVE_OUTPUT_DIRECTORY_DEBUG": "${sourceDir}/Binaries/${presetName}/Intermediate",
+                "CMAKE_C_COMPILER": "gcc",
+                "CMAKE_CXX_COMPILER": "g++"
+            }
+        },
+        {
+            "name": "Release",
+            "displayName": "Release Configuration",
+            "description": "Ninja build system is used for release.",
+            "generator": "Ninja",
+            "binaryDir": "${sourceDir}/Binaries/${presetName}/Intermediate",
+            "cacheVariables": {
+                "CMAKE_BUILD_TYPE": "Release",
+                "CMAKE_TOOLCHAIN_FILE": "",
+                "CMAKE_INSTALL_PREFIX": "${sourceDir}/Binaries/${presetName}/Output",
+                "CMAKE_ARCHIVE_OUTPUT_DIRECTORY_DEBUG": "${sourceDir}/Binaries/${presetName}/Intermediate"
+            }
+        }
+    ],
+    "buildPresets": [
+        {
+            "name": "Debug",
+            "description": "",
+            "displayName": "",
+            "configurePreset": "Debug"
+        },
+        {
+            "name": "Release",
+            "description": "",
+            "displayName": "",
+            "configurePreset": "Release"
+        }
+    ],
+    "testPresets": [
+        {
+            "name": "GoogleTest",
+            "description": "Google test framework",
+            "displayName": "GTest",
+            "configurePreset": "Debug"
+        }
+    ]
+}
