+++ Demo/Demo.cpp
@@ -17,7 +17,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-#include "dubins.h"
+#include "DubinsCurve/DubinsCurve.h"
 #include <stdio.h>
 
 int printConfiguration(double q[3], double x, void* user_data) {
